Issue BGDIINF_SB-1524 : monkey patch running in the background causes warning on startup and might have inconsistent comportment at runtime.
The issue is caused by the fact that the boto library was loading a ssl module (through urllib._ssl) before gevent monkey could patch the ssl module. To solve the issue, it is necessary to force the patch manually before importing boto.

Since boto is imported in the models, which are imported in the routes, which are imported in the app, which is imported in the wsgi.py (the first file loaded), we need to patch this at the very beginning of the file.

We also ran the linter in this commit.

disable pylint for wrong import position of gevent.monkey