BGDIINF_SB-2027: Set Access-Control-Allow-Methods header according to the routes
The Access-Control-Allow-Methods header value is set accordingly to the
method registered with the endpoint.

Also moved code out of app.helpers.__init__.py as it is not a best
practice.

The init_logging() is also removed from the app.__init__.py to avoid
configuring logging twice when started with gunicorn. In this case
gunicorn configure loggging. The init_logging() is added to
service_launcher in order to have logging configured when starting the
flask development server.