Check existence of full URL for a shortlink-id before redirecting to it
As redirect=true per default, a GET to /<shortlink_id> will redirect to the full URL.
Probably it might make sense to do a HEAD request to the full URL before redirecting to it.
If the URL exists, the service will redirect to it. Otherwise a 404 is returned with a proper
error message containing the full URL and the corresponding shortlink_id in it.
Not sure, if this would negatively influence performance too much,
though. To be tested.