Corrected the generic function get_registered_method()
The allowed method in CORS header where not correct, the list contained
only the methods for the endpoint (methods registered by the python
route function), instead of the methods registered to the path.

Because the service did not have any function registered to the same
path, the bug did not have any consequence, but in order to be correct and
avoid copy paste mistake the function has been corrected.

Also added the ALLOWED_DOMAINS in .env.default in order to simplify the
e2e test with localhost.