BGDIINF_SB-2380: Fixed collision and short ID length #major
The sort ID was initially (when produced by mf-chsdi3 project) using the
%x formatting (hexadecimal). This formatting has been forgotten during the
migration of the service into this micro-service repository. This had a negative
impact on the size of the short ID (size increased from 10 to 12 characters).

Also the implementation did not handle short id collisions as it was on mf-chsdi3.
On dynamodb the simplest collision setup is to use the primary key uniqueness
check (other attributes cannot be checked as unique). On the old implementation
on mf-chsdi3 this check was not done which resulted in overwrite in case of
collision. In the new service, because the primary key was not the shortlink_id
a collision resulted with duplicate shortlink_id in the table !

Now we use the shortlink_id as primary key and we make sure that collision
won't overwrite, but trigger retries instead. We allow up to 10 retry before
failing.

Also cleaned up the code with a lots of simplification.

!WARNING! This is a #major breaking changes as it requires a new DynamoDB table
having the shortlink_id attribute set as primary key.