BGDIINF_SB-3129: Fixed dockerfile and validators version
The validators packet v.020 implemented a new url parser which breaks url
paths with # characters. v.22 contains a fix, which does not solve the
problem yet. Master branch contains a newer fix which introduces the
strict_query=False parameter, but a release with this fix does not exist
so far. Thus, validators is pinned to v0.20 until a newer working version
exists.