PB-737: Fix CORS issues when working on localhost
Previous PR #66 was too restrictive has it blocked localhost on the DEV staging
which is needed to develop the web-mapviewer.

The main issue was that the ALLOWED_DOMAINS config was used for CORS and for
URL parameter check, but in the code the logic to validate CORS and url parameter
differ, due to this it was not possible anymore to set a ALLOWED_DOMAINS value
that would allow localhost as CORS and or as url parameter. The url parameter
check wanted to have a regex to exactly match the host name without the path,
while CORS did check the full url.

Now the code has been simplified and we use the same function/logic to test url
and CORS, using the full url. Note that ORIGIN and REFERER header based on the
documentation should always has the scheme http:// or https://. Referer might
have a path or not.

So now the ALLOWED_DOMAINS must be a pattern that match a full URL not only part
of it.